Switch 'Register Controller' panel to use Vanilla panel
