+++ src/components/Routes/Routes.js
@@ -15,7 +15,8 @@ export const paths = {
   "/": { redirect: "/models" },
   "/models": { component: Models },
   "/models/*": { component: ModelDetails },
-  "/controllers": { component: Controllers },
+  "/controllers/": { component: Controllers },
+  "/controllers/register": { component: Controllers },
   "/settings": { component: Settings },
 };
