+++ src/pages/Controllers/Controllers.js
@@ -1,8 +1,7 @@
 import { useState } from "react";
-import { useDispatch, useSelector, useStore } from "react-redux";
-import { Link } from "react-router-dom";
+import { useSelector } from "react-redux";
+import { Link, useHistory } from "react-router-dom";
 import cloneDeep from "clone-deep";
-import classNames from "classnames";
 
 import Layout from "components/Layout/Layout";
 import Header from "components/Header/Header";
@@ -11,18 +10,17 @@ import MainTable from "@canonical/react-components/dist/components/MainTable/Mai
 
 import FadeIn from "animations/FadeIn";
 
-import useLocalStorage from "hooks/useLocalStorage";
 import useWindowTitle from "hooks/useWindowTitle";
 
-import { getBakery, getControllerData, getModelData } from "app/selectors";
-import { connectAndStartPolling } from "app/actions";
+import { getControllerData, getModelData } from "app/selectors";
+
 import ControllersOverview from "./ControllerOverview/ControllerOverview";
 
 import "./_controllers.scss";
 
 function Details() {
   useWindowTitle("Controllers");
-
+  const history = useHistory();
   const controllerData = useSelector(getControllerData);
   const modelData = useSelector(getModelData);
   const [showRegisterAController, setShowRegisterAController] = useState(false);
@@ -90,7 +88,7 @@ function Details() {
       Registered
       <span
         className="controllers--registered-tooltip p-icon--help"
-        title="The controller authentication data is only stored in your browser localstorage. If you'd like this to persist across browsers try JAAS"
+        title="The controller authentication data is only stored in your browser localStorage. If you'd like this to persist across browsers try JAAS"
       ></span>
     </span>
   );
@@ -147,9 +145,7 @@ function Details() {
         <div className="controllers--register">
           <button
             className="p-button--positive"
-            onClick={() => {
-              setShowRegisterAController(!showRegisterAController);
-            }}
+            onClick={() => history.push("/controllers/register")}
           >
             Register a controller
           </button>
@@ -171,227 +167,12 @@ function Details() {
 }
 
 function RegisterAController({ onClose, showRegisterAController }) {
-  const [formValues, setFormValues] = useState({});
-  const dispatch = useDispatch();
-  const reduxStore = useStore();
-  const bakery = useSelector(getBakery);
-  const [additionalControllers, setAdditionalControllers] = useLocalStorage(
-    "additionalControllers",
-    []
-  );
-
-  function handleRegisterAController(e) {
-    e.preventDefault();
-    // XXX Validate form values
-    additionalControllers.push([
-      `wss://${formValues.wsControllerHost}/api`, // wsControllerURL
-      { user: formValues.username, password: formValues.password }, // credentials
-      null, // bakery
-      formValues.identityProviderAvailable, // identityProviderAvailable
-      true, // additional controller
-    ]);
-    setAdditionalControllers(additionalControllers);
-    dispatch(connectAndStartPolling(reduxStore, bakery));
-    onClose(); // Close the SlidePanel
-  }
-
-  function handleInputChange(e) {
-    const newFormValues = { ...formValues };
-    newFormValues[e.target.name] =
-      e.target.type === "checkbox" ? e.target.checked : e.target.value;
-    setFormValues(newFormValues);
-  }
-
-  function generateTheControllerLink(controllerIP) {
-    if (!controllerIP) {
-      return "the controller";
-    }
-    const dashboardLink = `https://${controllerIP}/dashboard`;
-    return (
-      <a href={dashboardLink} target="_blank" rel="noopener noreferrer">
-        the controller
-      </a>
-    );
-  }
-
   return (
     <SlidePanel
       onClose={onClose}
       isActive={showRegisterAController}
       className="register-a-controller__slide-panel"
-    >
-      <h5>Register a Controller</h5>
-      <p className="p-form-help-text">
-        Information can be retrieved using the <code>juju show-controller</code>{" "}
-        command.
-      </p>
-      <form
-        className="p-form p-form--stacked"
-        onSubmit={handleRegisterAController}
-      >
-        <div className="p-form__group row">
-          <div className="col-3">
-            <label
-              htmlFor="controller-name"
-              className="p-form__label is-required"
-            >
-              Name
-            </label>
-          </div>
-
-          <div className="col-8">
-            <div className="p-form__control">
-              <input
-                type="text"
-                id="controller-name"
-                name="controllerName"
-                onChange={handleInputChange}
-                required={true}
-              />
-              <p className="p-form-help-text">
-                Must be a valid alpha-numeric Juju controller name. <br />
-                e.g. production-controller-aws
-              </p>
-            </div>
-          </div>
-        </div>
-
-        <div className="p-form__group row">
-          <div className="col-3">
-            <label htmlFor="host" className="p-form__label is-required">
-              Host
-            </label>
-          </div>
-
-          <div className="col-8">
-            <div className="p-form__control">
-              <input
-                type="text"
-                id="host"
-                name="wsControllerHost"
-                onChange={handleInputChange}
-                required={true}
-              />
-              <p className="p-form-help-text">
-                You'll typically want to use the public IP:Port address for the
-                controller. <br />
-                e.g. 91.189.88.181:17070
-              </p>
-            </div>
-          </div>
-        </div>
-
-        <div className="p-form__group row">
-          <div className="col-3">
-            <label htmlFor="username" className="p-form__label">
-              Username
-            </label>
-          </div>
-
-          <div className="col-8">
-            <div className="p-form__control">
-              <input
-                type="text"
-                id="username"
-                name="username"
-                onChange={handleInputChange}
-              />
-              <p className="p-form-help-text">
-                The username you use to access the controller.
-              </p>
-            </div>
-          </div>
-        </div>
-
-        <div className="p-form__group row">
-          <div className="col-3">
-            <label htmlFor="password" className="p-form__label">
-              Password
-            </label>
-          </div>
-
-          <div className="col-8">
-            <div className="p-form__control">
-              <input
-                type="password"
-                id="password"
-                name="password"
-                onChange={handleInputChange}
-              />
-              <p className="p-form-help-text">
-                The password will be what you used when running{" "}
-                <code>juju register</code> or if unchanged from the default it
-                can be retrieved by running <code>juju dashboard</code>.
-              </p>
-            </div>
-          </div>
-        </div>
-
-        <div
-          className={classNames("p-form__group row", {
-            "u-hide": formValues.username && formValues.password,
-          })}
-        >
-          <div className="col-8 col-start-large-4">
-            <input
-              type="checkbox"
-              id="identityProviderAvailable"
-              name="identityProvider"
-              defaultChecked={false}
-              onChange={handleInputChange}
-            />
-            <label htmlFor="identityProviderAvailable">
-              An identity provider is available.{" "}
-            </label>
-            <div className="p-form-help-text identity-provider">
-              If you provided a username and password this should be left
-              unchecked.
-            </div>
-          </div>
-        </div>
-        <div className="row">
-          <div className="col-8 col-start-large-4">
-            <i className="p-icon--warning"></i>
-            <div className="controller-link-message">
-              Visit {generateTheControllerLink(formValues?.wsControllerHost)} to
-              accept the certificate on this controller to enable a secure
-              connection
-            </div>
-          </div>
-        </div>
-        <div className="row horizontal-rule">
-          <div className="col-8 col-start-large-4">
-            <input
-              type="checkbox"
-              id="certificateHasBeenAccepted"
-              name="certificateAccepted"
-              defaultChecked={false}
-              onChange={handleInputChange}
-              required={true}
-            />
-            <label htmlFor="certificateHasBeenAccepted">
-              The SSL certificate, if any, has been accepted.{" "}
-              <span className="required-star">*</span>
-            </label>
-          </div>
-        </div>
-        <div className="row register-a-controller__submit-segment push-1-rem">
-          <div className="col-12">
-            <button
-              className="p-button--positive u-float-right"
-              type="submit"
-              disabled={!formValues.certificateAccepted}
-            >
-              Add Controller
-            </button>
-            <p className="p-form-help-text">
-              The credentials are stored locally in your browser and can be
-              cleared on log-out.
-            </p>
-          </div>
-        </div>
-      </form>
-    </SlidePanel>
+    ></SlidePanel>
   );
 }
