+++ src/panels/_panels.scss
@@ -1,4 +1,17 @@
 @import "vanilla-framework/scss/vanilla";
+@import "./../scss/functions/z-index";
+
+.l-aside {
+  padding: 1rem;
+}
+
+.panel-container {
+  height: 100vh;
+  position: absolute;
+  right: 0;
+  top: 0;
+  z-index: z("infinity");
+}
 
 .panel {
   &-header {
